TabPanel TS conversion
